Crate bridge_types

Structs

ErrorStrings

Param

Struct used by sl_sh_fn macro to embed information in an array at runtime about each of the parameters of the rust native function.

Enums

PassingStyle

Used by sl_sh_fn macro to embed information at runtime about the parameters of the rust native function, specifically whether it is going to pass the value (a move), a reference, or mutable reference.

SloshChar

Type to hold Slosh’s notion of a char.

TypeHandle

Used by sl_sh_fn macro to embed information at runtime about the parameters of the rust native function, specifically whether it is a normal Type, or some supported wrapped type, e.g. Optional.

Traits

BridgedType

A slosh Value that can potentially be represented as a rust value. Marker traits

Type Aliases

LooseString

Type to hold anything in Slosh that can be represented as a String.

VarArgs

Public type used by rust native -> slosh bridge macro to represent arguments in slosh that correspond to variadic functions in rust