Crate bridge_types

Structs

ErrorStrings

Keyword

Type to hold anything in Slosh that can be represented as a slosh Keyword: u32.

KeywordAsString

Type to hold anything in Slosh that can be represented as a slosh symbol, however, this type, in contrast to Keyword does need to do some allocations, specifically when using [SlFrom]. Not a huge penalty but worth mentioning if unnecessary.

LooseFloat

Type to hold anything in Slosh that can be represented as a slosh float value: [u8; 7].

LooseInt

Type to hold anything in Slosh that can be represented as slosh int value: [u8; 7].

Param

Struct used by sl_sh_fn macro to embed information in an array at runtime about each of the parameters of the rust native function.

Symbol

Type to hold anything in Slosh that can be represented as a slosh symbol: u32.

SymbolAsString

Type to hold anything in Slosh that can be represented as a slosh symbol, however, this type, in contrast to Symbol does need to do some allocations, specifically when using [SlFrom]. Not a huge penalty but worth mentioning if unnecessary.

Enums

PassingStyle

Used by sl_sh_fn macro to embed information at runtime about the parameters of the rust native function, specifically whether it is going to pass the value (a move), a reference, or mutable reference.

SloshChar

Type to hold Slosh’s notion of a char.

TypeHandle

Used by sl_sh_fn macro to embed information at runtime about the parameters of the rust native function, specifically whether it is a normal Type, or some supported wrapped type, e.g. Optional.

Traits

BridgedType

A slosh Value that can potentially be represented as a rust value. Marker traits

Type Aliases

LooseString

Type to hold anything in Slosh that can be represented as a String.

VarArgs

Public type used by rust native -> slosh bridge macro to represent arguments in slosh that correspond to variadic functions in rust